pyproject.toml
[project]
name = "lithic-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "monty@zenitsu" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "odmantic>=1.0.2",
    "openai>=1.101.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
]

[project.scripts]
lithic-api = "lithic_api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
]
